+++ packages/terra-form/tests/nightwatch/select/select-spec.js
@@ -1,19 +1,10 @@
-/* eslint import/no-extraneous-dependencies: ["error", {"devDependencies": true}] */
 // eslint-disable-next-line import/no-extraneous-dependencies
-const screenshot = require('terra-toolkit').screenshot;
+const resizeTo = require('terra-toolkit/lib/nightwatch/responsive-helpers').resizeTo;
 
-module.exports = {
-  before: (browser, done) => {
-    browser.resizeWindow(browser.globals.width, browser.globals.height, done);
-  },
-
-  afterEach: (browser, done) => {
-    screenshot(browser, 'terra-form', done);
-  },
-
+module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous'], {
   'Displays a default Select with a single option': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/form-tests/select/default`)
+      .url(`${browser.launchUrl}/#/tests/form-tests/select/default`)
       .assert.elementPresent('select')
       .assert.elementPresent('option:nth-of-type(1)') // 1 option present
       .assert.elementNotPresent('option:nth-of-type(2)');
@@ -21,7 +12,7 @@ module.exports = {
 
   'Displays a Select with a single option, using deprecated choices prop': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/form-tests/select/choices_prop`)
+      .url(`${browser.launchUrl}/#/tests/form-tests/select/choices_prop`)
       .assert.elementPresent('select')
       .assert.elementPresent('option:nth-of-type(1)') // 1 option present
       .assert.elementNotPresent('option:nth-of-type(2)');
@@ -29,7 +20,7 @@ module.exports = {
 
   'Displays a populated Select with multiple options': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/form-tests/select/populated`)
+      .url(`${browser.launchUrl}/#/tests/form-tests/select/populated`)
       .assert.elementPresent('select')
       .assert.elementPresent('option:nth-of-type(2)') // 2 options present
       .assert.elementNotPresent('option:nth-of-type(3)');
@@ -37,7 +28,7 @@ module.exports = {
 
   'Displays a populated Select with a defaultValue selected and can change selection': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/form-tests/select/populated`)
+      .url(`${browser.launchUrl}/#/tests/form-tests/select/populated`)
       .assert.value('select[name="zounds"]', 'b')
       .assert.containsText('select[name="zounds"]', 'boo')
       .setValue('select[name="zounds"]', 'm')
@@ -47,7 +38,7 @@ module.exports = {
 
   'Displays a populated Select with a onChange and value for controlled form submission': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/form-tests/select/populated_controlled`)
+      .url(`${browser.launchUrl}/#/tests/form-tests/select/populated_controlled`)
       .assert.value('select[name="animal"]', 'micros')
       .assert.elementPresent('option:nth-of-type(4)') // 4 options present
       .assert.elementNotPresent('option:nth-of-type(5)');
@@ -55,10 +46,10 @@ module.exports = {
 
   'Controlled Select correctly submits form data': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/form-tests/select/populated_controlled`)
+      .url(`${browser.launchUrl}/#/tests/form-tests/select/populated_controlled`)
       .assert.value('select[name="animal"]', 'micros')
       .setValue('select[name="animal"]', 'kittens')
       .click('#select-submit')
       .assert.containsText('#select-submission', 'Submitted with data {"animal":"kittens"}');
   },
-};
+});