+++ packages/terra-overlay/tests/nightwatch/OverlayContainer/overlay-container-spec.js
@@ -1,18 +1,9 @@
-/* eslint import/no-extraneous-dependencies: ["error", {"devDependencies": true}] */
 // eslint-disable-next-line import/no-extraneous-dependencies
-const screenshot = require('terra-toolkit').screenshot;
+const resizeTo = require('terra-toolkit/lib/nightwatch/responsive-helpers').resizeTo;
 
-module.exports = {
-  before: (browser, done) => {
-    browser.resizeWindow(browser.globals.width, browser.globals.height, done);
-  },
-
-  afterEach: (browser, done) => {
-    screenshot(browser, 'terra-overlaycontainer', done);
-  },
-
+module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous'], {
   'Displays a default OverlayContainer': (browser) => {
-    browser.url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/overlay-container-tests/default`);
+    browser.url(`${browser.launchUrl}/#/tests/overlay-container-tests/default`);
     browser.assert.attributeEquals('#terra-OverlayContainer', 'tabIndex', '-1');
   },
-};
+});