+++ packages/terra-base/CHANGELOG.md
@@ -3,6 +3,11 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update nightwatch tests and test scripts.
+
+2.0.0 - (August 31, 2017)
+------------------
 ### Added
 * Added `--terra-base-line-height` CSS custom property
 * Added `--terra-base-html-height` CSS custom property
@@ -15,7 +20,7 @@ Unreleased
 
 ### Removed
 * Removed global border, margin, and padding resets from base styles
-* Removed global list-style reset from ol and ul elements
+* Removed global list-style reset from `ol` and `ul` elements
 
 1.4.0 - (August 15, 2017)
 ------------------