+++ packages/terra-image/CHANGELOG.md
@@ -3,8 +3,14 @@ Changelog
 
 Unreleased
 ----------
+
+1.5.0 - (August 31, 2017)
+------------------
 ### Added
 * Added default border, margin, and padding reset to img
+
+### Changed
+* Update nightwatch tests and test scripts.
 
 1.4.0 - (August 15, 2017)
 ------------------