+++ packages/terra-time-input/CHANGELOG.md
@@ -6,6 +6,15 @@ Unreleased
 ### Fixed
 * Fixed focus management issue between hour and time input in Firefox.
 
+### Changed
+* Update nightwatch tests and test scripts.
+
+1.6.0 - (August 31, 2017)
+------------------
+### Changed
+* Minor version bump
+
+
 1.5.0 - (August 15, 2017)
 ------------------
 ### Changed